% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllCorrelations.R
\name{AllCorrelations}
\alias{AllCorrelations}
\title{Correlations between opportunities for selection and population growth rates}
\usage{
AllCorrelations(Opportunities, PopulationGrowth)
}
\arguments{
\item{Opportunities}{A list created by the function Fiz()}

\item{PopulationGrowth}{A table created by the function PopGrowthFromID()}
}
\value{
A data frame with correlations and their standard errors for the different types of opportunity for selection
}
\description{
Calculate the year-based correlations between the various opportunities for selection (total, and trait-specific), and population growth rates
}
\examples{
data("svdata")

Opportunities <- Fiz(dat = svdata, fitnesstrait = "FitnessTot", 
phenotraits = c("Mass", "BL", "TL"), yearname = "Year", 
covariates = c("Sex", "Age"), family = "poisson")
PopulationGrowth <- PopGrowthFromID(dat=svdata, idname="ID", yearname="Year")  
AllCorrelations(Opportunities = Opportunities, PopulationGrowth = PopulationGrowth)

}
\seealso{
\code{\link{Fiz}}, \code{\link{PopGrowthFromID}} and \code{\link{cor_with_se}}
}
