% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalyseAll.R
\name{AnalyseAll}
\alias{AnalyseAll}
\title{Analyses the correlation between the opportunity for selection and population growth from raw  (individual level) data}
\usage{
AnalyseAll(dat, fitnesstrait, phenotraits, idname = "ID", yearname = "Year",
  MinDataPoints = 5, covariates = NULL, NAsToMean = TRUE,
  glm_family = "gaussian", survivalname = NULL, insertmissing = FALSE,
  outputfile = NULL, SummaryStatsfiles = NULL)
}
\arguments{
\item{dat}{Data-frame. Contains all necessary data.}

\item{fitnesstrait}{Character vector (length>=1). Name of the column(s) containing fitness data.}

\item{phenotraits}{Character vector (length>=1). Name of the column(s) containing phenotypic data.}

\item{idname}{Character (length=1). Name of the column containing the individual unique identifiers}

\item{yearname}{Character (length=1). Name of the column containing the time-step identity (typically year). Default="Year".}

\item{MinDataPoints}{Integer. Minimal number of data points for a trait on one year. If there are fewer data, the trait is not considered for that year. Default=5.}

\item{covariates}{NULL of character vector (length>=1). Name of the column(s) containing covariates which effect should be removed from calculations. Default=NULL.}

\item{NAsToMean}{Boolean. Should missing phenotypic values be changed to the trait mean. Default=TRUE.}

\item{glm_family}{Character vector. Family of the (G)LMs to be fitted. Possible values are "gaussian", "poisson", "binomial". Default="gaussian". If there is more than one fitness measure considered this should be the same length as the number of fitness measures. If only one family is passed, calculations for all fitness measures are assumed to follow this distribution.}

\item{survivalname}{Character (length=1). Name of the column containing the survival data. Defaults to NULL. Does not need to be changed from default if insertmissing=FALSE}

\item{insertmissing}{boolean. Should missing individual data be added (individuals not present in a year but known to be in the population as they are present in a subsequent year). Defaults to FALSE}

\item{outputfile}{Character (length=1). File path that the resulting data should be written to, if so required. Defaults to NULL (no file is saved).}

\item{SummaryStatsfiles}{Name of a csv file with summary statistics for opportunity for selection, for all fitness traits, and population growth rate. If NULL (default) does not write it.}
}
\value{
A data frame with correlations and their standard errors for the different types of opportunity for selection for each of the fitness measures passed to the function.
}
\description{
Starting from individual level data, first calculates population growth rates and opportunity for selection in each year (and for each fitness and phenotypic trait specified). The correlation between these measures is then calculated.
}
\examples{
data("svdata")

AnalyseAll(svdata, fitnesstrait=c("FitnessTot","Phi"), phenotraits=c("Mass","BL","TL"), 
idname="ID", yearname="Year", glm_family =c("gaussian", "binomial"),
 insertmissing=TRUE, survivalname=NULL)

}
\seealso{
\code{\link{Fiz}}, \code{\link{PopGrowthFromID}}, \code{\link{cor_with_se}}, \code{\link{InsertMissing}} and \code{\link{AllCorrelations}}
}
