% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_with_se.R
\name{cor_with_se}
\alias{cor_with_se}
\title{Correlation with standard error}
\usage{
cor_with_se(x, y, wt = NULL)
}
\arguments{
\item{x}{Numeric vector.}

\item{y}{Numeric vector.}

\item{wt}{Numeric vector or NULL. If not NULL, contains weights.}
}
\value{
A vector of length 2, containing the estimated correlation between x and y, followed by the standard error of the estimate.
}
\description{
Calculates correlation between x and y, possibly with weights, with standard error.
}
\examples{
x <- rnorm(100)
y <- rnorm(100) + x
cor_with_se(x,y)

}
