\name{coropgro-package}
\alias{coropgro-package}
\alias{coropgro}
\docType{package}
\title{
\packageTitle{coropgro}
}
\description{
\packageDescription{coropgro}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{coropgro}
\packageIndices{coropgro}
Check the vignette for a demonstration.
}
\author{
\packageAuthor{coropgro}

Maintainer: \packageMaintainer{coropgro}
}
\references{
NA
}
\keyword{ package }
\examples{
library(coropgro)
data("svdata")
}
