% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fiz.R
\name{Fiz}
\alias{Fiz}
\title{Calculates trait-specific opportunity for selection}
\usage{
Fiz(dat, fitnesstrait, phenotraits, yearname = "Year", MinDataPoints = 5,
  covariates = NULL, NAsToMean = TRUE, family = "gaussian")
}
\arguments{
\item{dat}{Data-frame. Contains all necessary data.}

\item{fitnesstrait}{Character (length=1). Name of the column containing fitness data.}

\item{phenotraits}{Character vector (length>=1). Name of the column(s) containing phenotypic data.}

\item{yearname}{Character (length=1). Name of the column containing the time-step identity (typically year). Default="Year".}

\item{MinDataPoints}{Integer. Minimal number of data points for a trait on one year. If there are fewer data, the trait is not considered for that year. Default=5.}

\item{covariates}{NULL of character vector (length>=1). Name of the column(s) containing covariates which effect should be removed from calculations. Default=NULL.}

\item{NAsToMean}{Boolean. Should missing phenotypic values be changed to the trait mean. Default=TRUE.}

\item{family}{Character. Family of the (G)LMs to be fitted. Possible values are "gaussian", "poisson", "binomial". Default="gaussian".}
}
\value{
A list.
}
\description{
Calculates trait-specific opportunity for selection
}
\examples{
data("svdata")
Fiz(dat=svdata, fitnesstrait="Phi", phenotraits=c("Mass", "BL", "TL"), family="binomial")

}
