% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InsertMissing.R
\name{InsertMissing}
\alias{InsertMissing}
\title{Inserts individuals missing in years they were alive but not measured}
\usage{
InsertMissing(data, idname = "ID", yearname = "Year",
  survivalname = "survival", verbose = FALSE)
}
\arguments{
\item{data}{- data frame containing individual data on fitness}

\item{idname}{The name of the column containing individual identification, defaults to "ID"}

\item{yearname}{The name of the column containing year identity (or other time steps), defaults to "Year"}

\item{survivalname}{The name of the column containing survival information, defaults to "survival"}

\item{verbose}{Boolean. If TRUE, print the number of missing entries added}
}
\value{
A data frame containing all the original data, with missing data added in. If no missing data was found, the original dataframe is returned.
}
\description{
Find individuals that are present in multiple years, but missing from the phenotypic data when they were known to be alive. Gives those individuals data for survival (1, as they survived) and missing data for fecundity.
}
\examples{
data("svdata")
svdata <- InsertMissing(data = svdata, idname = "ID", yearname = "Year", 
survivalname = "Phi", verbose = TRUE)

}
