% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpportunitiesCSV.R
\name{OpportunitiesCSV}
\alias{OpportunitiesCSV}
\title{Reformats data on the opportunity for selection}
\usage{
OpportunitiesCSV(Opportunities, lambda = NULL, outputfile = NULL)
}
\arguments{
\item{Opportunities}{A list created by the function Fiz()}

\item{lambda}{A vector of population growth rates created by the function PopGrowthFromID(), if that is to be included in the file. Defaults to NULL}

\item{outputfile}{File path for where the reformatted dataframe should be saved to. Default is NULL (no file written)}
}
\value{
A long form data frame containing information on the opportunity for selection over multiple years.
}
\description{
Takes data produced by the Fiz() function and reformats in a long-table to be output to a csv file. Optionally can write direct to file
}
\examples{
data("svdata")

Opportunities <- Fiz(dat = svdata, fitnesstrait = "FitnessTot", 
phenotraits = c("Mass", "BL", "TL"), yearname = "Year", 
covariates = c("Sex", "Age"), family = "poisson")
PopulationGrowth <- PopGrowthFromID(dat=svdata, idname="ID", yearname="Year")  
OpportunitiesCSV(Opportunities, lambda=PopulationGrowth, outputfile="Snowvoleopportunities.csv")

}
\seealso{
\code{\link{Fiz}}
}
