% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PopGrowthFromID.R
\name{PopGrowthFromID}
\alias{PopGrowthFromID}
\title{Calculate population growth rate between years}
\usage{
PopGrowthFromID(dat, idname = "ID", yearname = "Year")
}
\arguments{
\item{dat}{Data-frame. Contains all necessary data.}

\item{idname}{Column name containing individual unique identifiers}

\item{yearname}{Column name containing years (or other time step). 

#' @return A table of annual population growth rate.

#' @examples
data("svdata")
PopGrowthFromID(dat=svdata, idname="ID", yearname="Year")}
}
\description{
Uses minimal and maximal individual presence to compute changes in population size.
}
