% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fiz.R
\docType{data}
\name{svdata}
\alias{svdata}
\title{Snow vole data}
\format{A data frame with 1423 rows and 10 variables}
\usage{
data(svdata)
}
\description{
A dataset containing individual-based phenotypic data from a snow vole population. The variables are as follows:
}
\details{
\itemize{
  \item ID Individual unique identifier
  \item Year Year of measurement
  \item Sex Factor, "Female" or "Male"
  \item Age Factor, "A" or "J"
  \item Mass Body mass in grams
  \item BL Body length in cm
  \item TL Tail length in cm
  \item Phi Boolean, survival to the next year
  \item RhoZ Reproductive success on the next year
  \item FitnessTot Sum of RhoZ and Phi
}
}
\keyword{datasets}
