% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WriteSummaryStats.R
\name{WriteSummaryStats}
\alias{WriteSummaryStats}
\title{Write opportunities and population growth to csv file}
\usage{
WriteSummaryStats(Opportunities, PopulationGrowth,
  FileName = "SummaryStatsIpop.csv", write = TRUE, fitnesstrait = NA,
  family = NA)
}
\arguments{
\item{Opportunities}{Output from a call to Fiz().}

\item{PopulationGrowth}{Output from a call to PopGrowthFromID().}

\item{FileName}{Name of the file where the output sould be written (including directory if not to be printed in the working directory).}

\item{write}{Boolean. If TRUE, write the file. If FALSE, return the output instead.}

\item{fitnesstrait}{Name of the fitness trait used for opportunity for selection.}

\item{family}{GLM family ("gaussian", "poisson" or "binomial")}
}
\value{
Write to a file or return a data frame
}
\description{
Write year-specific opportunities for selection and population growth rate to a csv file
}
\examples{

data(svdata)
Opportunities <- Fiz(dat=svdata, fitnesstrait="Phi",
 phenotraits=c("Mass", "BL", "TL"), family="binomial")
PopulationGrowth <- PopGrowthFromID(dat=svdata, idname="ID", yearname="Year")

WriteSummaryStats(Opportunities=Opportunities, PopulationGrowth=PopulationGrowth )

}
\seealso{
\code{\link{Fiz}} and \code{\link{PopGrowthFromID}}
}
